+++ CHANGES.md
@@ -8,6 +8,7 @@ Change Log
 * Added support for `parameters` to `ArcGisMapServerImageryProvider`.
 * Hacky fix/workaround for https://github.com/AnalyticalGraphicsInc/cesium/issues/7121.
 * Added support for using the imagery splitting system with the terrain and the atmosphere, too.
+* Added support for extruding `GeoJsonDataSource` polygons using the `heightReference`, `heightProperty`, and `extrudedHeightProperty` properties.
 
 ### 1.50 - 2018-10-01